Add files via upload
This is my second assignment since joining the college, which mainly includes the project code and the explanation of important knowledge points. There is a. txt text format file that contains the entire process of my proof of key knowledge. It is a webpage that can be copied and viewed.